requirement.txt
jieba
pytorch
